# api/docs/swagger.yaml
basePath: /
definitions:
  api.apiError:
    properties:
      code:
        type: string
      message:
        type: string
      param:
        type: string
      type:
        type: string
    type: object
  api.apiErrorResponse:
    properties:
      error:
        $ref: '#/definitions/api.apiError'
      object:
        description: '"error"'
        type: string
    type: object
...
  api.deletedResponse:
    properties:
      deleted:
        type: boolean
      id:
        type: string
      object:
        type: string
    type: object
...
  api.listResponse-api_sessionResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/api.sessionResponse'
        type: array
      has_more:
        type: boolean
      next_cursor:
        type: string
      object:
        description: '"list"'
        type: string
    type: object
...
  api.loginRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: secret
        type: string
    type: object
...
  api.refreshRequest:
    properties:
      refresh_token:
        example: a3f1...
        type: string
    type: object
  api.registerRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: secret
        type: string
    type: object
  api.sessionResponse:
    properties:
      created:
        type: integer
      expires_at:
        type: integer
      id:
        type: string
      object:
        description: '"session"'
        type: string
      user_id:
        type: string
    type: object
  api.tokenResponse:
    properties:
      access_token:
        type: string
      object:
        description: '"token"'
        type: string
      refresh_token:
        type: string
    type: object
  api.updateUserRequest:
    properties:
      email:
        example: new@example.com
        type: string
      role:
        example: admin
        type: string
    type: object
  api.userResponse:
    properties:
      created:
        type: integer
      email:
        type: string
      id:
        type: string
      object:
        description: '"user"'
        type: string
      role:
        type: string
    type: object
...
host: legist.nadevko.cc
info:
  contact: {}
  description: |-
    AI assistant for comparing revisions of normative legal acts (НПА) of the Republic of Belarus.

    ## Authentication
    All endpoints except public auth routes require a Bearer token in the `Authorization` header.
    Obtain tokens via `POST /sessions`. Refresh via `POST /tokens/refresh`.

    ## Request IDs
    Every response includes a `Request-Id` header for debugging.

    ## Idempotency
    `POST` requests require an `Idempotency-Key` header.
    Repeating a request with the same key returns the cached response.
    Keys expire after 24 hours. Reusing a key for a different endpoint returns 422.

    ## Roles
    Users have `role`: `user` or `admin` (in `users` table and JWT `role` claim).
    Only `admin` may change another user's role via `PATCH /users/:id`.
    In dev (`ENV=dev`), new registrations default to `admin`; otherwise default is `user`.

    ## Ownership and access control
    Resources belong to the authenticated user. Attempting to read or mutate
    another user's private resource returns 404 (not 403) to avoid leaking existence.
    Public resources (`user_id IS NULL`) are readable by any authenticated user but cannot be mutated.
    List `owner` query (Stripe-like): omit → non-admin sees only own rows; admin sees own + public.
    `owner=null` or `owner=public` (aliases, admin only) → public rows only. `owner=<your user id>` → own rows only. Other ids are rejected.

...

...

...

...

...

...

...

...

...

...
paths:
...
  /sessions:
    get:
      parameters:
      - description: Limit
        in: query
        name: limit
        type: integer
      - description: Cursor
        in: query
        name: starting_after
        type: string
      - description: Cursor
        in: query
        name: ending_before
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.listResponse-api_sessionResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.apiErrorResponse'
      security:
      - BearerAuth: []
      summary: List active sessions
      tags:
      - Sessions
    post:
      consumes:
      - application/json
      parameters:
      - description: Email and password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.tokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.apiErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.apiErrorResponse'
      summary: Login
      tags:
      - Sessions
  /sessions/{id}:
    delete:
      parameters:
      - description: Session ID
        in: path
        name: id
        required: true
        type: string
      - description: Idempotency key
        in: header
        name: Idempotency-Key
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.deletedResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.apiErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.apiErrorResponse'
      security:
      - BearerAuth: []
      summary: Logout (delete session)
      tags:
      - Sessions
...
  /tokens/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: Refresh token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.refreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.tokenResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.apiErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.apiErrorResponse'
      summary: Refresh access token
      tags:
      - Sessions
  /users:
    post:
      consumes:
      - application/json
      parameters:
      - description: Email and password
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.registerRequest'
      - description: Idempotency key
        in: header
        name: Idempotency-Key
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/api.userResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.apiErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/api.apiErrorResponse'
      summary: Register a new user
      tags:
      - Sessions
  /users/{id}:
    delete:
      parameters:
      - description: User ID or 'me'
        in: path
        name: id
        required: true
        type: string
      - description: Idempotency key
        in: header
        name: Idempotency-Key
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.deletedResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/api.apiErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.apiErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - Users
    get:
      parameters:
      - description: User ID or 'me'
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.userResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.apiErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.apiErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user
      tags:
      - Users
    patch:
      consumes:
      - application/json
      parameters:
      - description: User ID or 'me'
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update (email and/or role)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.updateUserRequest'
      - description: Idempotency key
        in: header
        name: Idempotency-Key
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.userResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.apiErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/api.apiErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.apiErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/api.apiErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user (email for self; role for admins)
      tags:
      - Users
...
securityDefinitions:
  BearerAuth:
    description: 'Enter token in format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...